+++ CHANGELOG.md
@@ -54,6 +54,9 @@
 - `WeldxFile.show_asdf_header` prints output on console, before it only returned the header as parsed dict and string
   representation. Also tweaked efficiency by not writing binary
   blocks [[#459]](https://github.com/BAMWelDX/weldx/pull/459), [[#469]](https://github.com/BAMWelDX/weldx/pull/469).
+- Merging and unmerging multiple `CoordinateSystemManager` instances now correctly preserves all attached data.
+  [[#494]](https://github.com/BAMWelDX/weldx/pull/494).
+- `compare_nested` can compare sets [[#496]](https://github.com/BAMWelDX/weldx/pull/496)
 
 ### documentation